Working on debugging
The edge case in which both temporal locations have the same timestamp is currently not being handled appropriately. Simply incrementing one history randomly leads to the possibility of missing an occurrence, and hence a contact. Given the application of this algorithm, this is not acceptable behavior. Going to refactor so that we recurse each time this occurs to consider both cases of incrementing just one history randomly. Given that this is an edge case and histories won't be very long, we shouldn't need to worry about stack overflows.

Also added loggging_config and the evaluation package.