Checkpoint
Implements heuristic: only send a factor message if its value or time is greater than the previous message sent to that variable. It is not enough to just check the value since the time influences the weight of the message. While this effectively requires storing another graph, it should reduce the number of ineffective messages sent significantly.

Implements graph partitioning using pymetis