Revert to clip(); function call optimizations; name changes
Clipping is necessary since the filtering is done with a buffer, so the time diffs could be greater than 0.
Oversaw a couple method calls that were still inside for loops.
Renamed fkey() to ckey() to indicate that it is for a contact.
Renamed _results() to _map() to slightly less confusing.