Checkpoint
I believe the issue with risk propagation lies in the partitioning algorithm. I tested this by creating a population out of a Gaussian mixture with two components. Yet, METIS does a poor job of cleanly partitioning this graph. The graph is disconnected which may be causing METIS to perform poorly. I'm going to try evaluating with generated graphs using NetworkX. I am also going to refactor the synthetic data generation to be more modular so that it's easier to swap out how data is generated.