Remove private attributes; overlap computation
Refactor to avoid storing stateful private attributes in RiskPropagation, and instead require them as explicit inputs to the methods that need them.
Refactor _init so that the exposure scores of users without neighbors are computed while message passing is active. In the case that there are a high number of disconnected nodes, this will save some time.