Bug fixes
Remove the filter in ContactFactory to check if the names are distinct.
Add early stopping to parameter and scalability experiments. This is particularly for the parameter experiments where very high transmission probability and very low send tolerance were resulting in hanging behavior.
Add graph simplification to remove self loops. This was previously an error.

Signed-off-by: rtatton <rdt17@protonmail.com>